Added supper class for all message status classes
